# GH-3957 split NOT IN and NOT EXISTS into separate tokens