… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/CharStream.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/JJTSyntaxTreeBuilderState.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/ParseException.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/SyntaxTreeBuilder.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/SyntaxTreeBuilderConstants.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/SyntaxTreeBuilderDefaultVisitor.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/SyntaxTreeBuilderTokenManager.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/SyntaxTreeBuilderTreeConstants.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/SyntaxTreeBuilderVisitor.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/Token.java
… core/queryparser/sparql/src/main/java/org/eclipse/rdf4j/query/parser/sparql/ast/TokenMgrError.java