GitHub issue resolved: #3957  <!-- add a Github issue number here, e.g #123. -->

Briefly describe the changes proposed in this PR:

- in sparql JavaCC grammar, split `NOT_IN` and `NOT_EXISTS` into separate tokens to allow arbitrary whitespace between
- regenerated SPARQL parser from grammar using JavaCC 7.0.4
- added two regression tests

----
PR Author Checklist (see the [contributor guidelines](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md) for more details):

 - [x] my pull request is [self-contained](https://rdf4j.org/documentation/developer/merge-strategy/#self-contained-changes-pull-requests-and-commits)
 - [x] I've added tests for the changes I made
 - [x] I've applied [code formatting](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md#code-formatting) (you can use `mvn process-resources` to format from the command line)
 - [x] I've [squashed](https://rdf4j.org/documentation/developer/squashing) my commits where necessary 
 - [x] every commit message starts with the issue number (GH-xxxx) followed by a meaningful description of the change