@hmottestad 2022-06-06
I remember I split up one of the very long methods somewhere in the generated code to help the compiler. I need to check if that is still an issue now.