@hmottestad 2022-06-06
## Before

## After
```
Benchmark                                                     Mode  Cnt      Score     Error  Units
QueryBenchmark.complexQuery                                   avgt    5      4.450 ±   0.032  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5      2.332 ±   0.041  ms/op
QueryBenchmark.groupByQuery                                   avgt    5      1.106 ±   0.032  ms/op
QueryBenchmark.long_chain                                     avgt    5    701.260 ±   3.767  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    193.124 ±   0.578  ms/op
QueryBenchmark.minus                                          avgt    5  15395.243 ± 173.074  ms/op
QueryBenchmark.nested_optionals                               avgt    5    204.214 ±   1.559  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     17.498 ±   0.045  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5      1.333 ±   0.072  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5     87.120 ±   0.620  ms/op
QueryBenchmark.simple_filter_not                              avgt    5      9.477 ±   0.192  ms/op
```