@hmottestad 2022-06-06
It's about 3x slower. I think it's probably the JVM compiling issue we've run into previously.