@hmottestad 2022-06-06
## After splitting up the big SyntaxTreeBuilderTokenManager#jjMoveNfa_0 method

```
Benchmark                                        Mode  Cnt  Score   Error  Units
SPARQLParseBenchmark.complexPathExpressionQuery  avgt    5  8.564 ± 0.857  ms/op
SPARQLParseBenchmark.selectValuesQuery           avgt    5  7.515 ± 0.403  ms/op
```