@jeenbroekstra 2022-06-06
> ## After splitting up the big SyntaxTreeBuilderTokenManager#jjMoveNfa_0 method
> 
> ```
> Benchmark                                        Mode  Cnt  Score   Error  Units
> SPARQLParseBenchmark.complexPathExpressionQuery  avgt    5  8.564 ± 0.857  ms/op
> SPARQLParseBenchmark.selectValuesQuery           avgt    5  7.515 ± 0.403  ms/op
> ```

We should do several things:

- create a patch file with this specific fix, so that we can reapply on the generated code after rerunning javacc. I suggest separate patch files for each specific fix.
- add info in a readme file on the how/why of the fix. This is necessary to help fix if the patch file no longer applies without problems (e.g when we bump javacc versions).