@hmottestad 2022-06-06
We have an existing patch file for this actually.

Can I propose we create some sort of script that runs JavaCC and applies all the patch files?

Then we could have a nice folder structure:
```
javaCC-sparql-generator
    patches
        large-method-jit
            README.md
            fix.patch
    script.sh
```