@hmottestad 2022-06-06
I've started on an issue for this here: https://github.com/eclipse/rdf4j/issues/3962

In the mean time I would like to merge this PR. I also have a[ fix the the serializable issue](https://github.com/eclipse/rdf4j/pull/3963) with SimpleLiteral, if you could review that for me then I will start on releasing 4.0.2 either tonight or tomorrow.