@jeenbroekstra 2022-06-06
> We have an existing patch file for this actually.

Yes but none of it works, presumably because the previous version of the parser was generated with javacc 5.0. And because none of it is documented or motivated I have no idea how to update that patch (or even which parts of it are still relevant). That patch file should be deleted/replaced.

> 
> Can I propose we create some sort of script that runs JavaCC and applies all the patch files?

I don't think that's a good idea. I'd rather just document the steps. It's not fully automated, there's some inspection and adjustment involved in some steps. 

> 
> Then we could have a nice folder structure:
> ```
> javaCC-sparql-generator
>     patches
>         large-method-jit
>             README.md
>             fix.patch
>     script.sh
> ```
>